other: organize some utility function files
